[refactor] multi module projects and update docs
