+++ boxer-binlog/src/test/resources/delta-streamer-config/hoodie_table_meta.properties
