+++ docs/document/02.Elasticsearch Ingestion.md
@@ -1,0 +1,1 @@
+# Document Data Ingestion via Elasticsearch
